delete webhook message

+++ src/index.js
@@ -10,7 +10,7 @@ const app = express()
 app.use(express.json())
 app.use(FileUpload())
 
-const deta = new Deta()
+const deta = new Deta(process.env.DETA_PROJECT_KEY)
 const db = deta.Base("data")
 
 app.get("/", async (req, res) => {
@@ -26,7 +26,7 @@ app.post("/file/upload", async (req, res) => {
     if (req.body["api_key"] !== (await db.get("_API_KEY"))?.value) return res.sendStatus(403)
     if (!req.files?.file) return res.sendStatus(400).send("No files were uploaded")
     const file = req.files.file
-    if (!file.size > 26214400) return res.sendStatus(413) // up to 25mb
+    if (file.size > 26214400) return res.sendStatus(413) // up to 25mb
 
     const code = generate()
     const form = new FormData()
@@ -37,17 +37,19 @@ app.post("/file/upload", async (req, res) => {
         body: form
     })
     if (!webhook.ok) return res.sendStatus(500)
-    const attachment = (await webhook.json()).attachments[0]
+    const json = await webhook.json()
     const uuid = generateUuid()
 
     await db.put({
-        url: attachment.url,
+        url: json.attachments[0].url,
+        id: json.id,
         deleteCode: uuid
     }, code)
 
     return res.send({
         code: code,
         url: `https://${req.headers.host}/${code}`,
+        discord_url: json.attachments[0].url,
         del_url: `https://${req.headers.host}/delete/${code}/${uuid}`
     })
 })
@@ -62,6 +64,10 @@ app.get("/delete/:code/:uuid", async (req, res) => {
     const file = await db.get(req.params.code)
     if (!file) return res.sendStatus(404)
     if (file.deleteCode !== req.params.uuid) return res.sendStatus(403)
+
+    const webhook = await fetch(`${process.env.DISCORD_WEBHOOK}/messages/${file.id}${process.env.THREAD_ID ? `?thread_id=${process.env.THREAD_ID}` : ""}`, { method: "DELETE" })
+    if (!webhook.ok) return res.sendStatus(500)
+
     await db.delete(req.params.code)
     return res.sendStatus(200)
 })
